finally fix bounds!

+++ 01_hello_world/script.js
@@ -7,8 +7,8 @@ window.onload = function() {
             var string =
                 new TextDecoder("utf8").decode(new Uint8Array(buffer, 0, 13));
             console.log(string);
-            console.log(new Uint32Array(buffer.slice(13, 17)));
+            console.log(new Float32Array(buffer.slice(13, 17)));
             object.instance.exports.main();
-            console.log(new Uint32Array(buffer.slice(13, 17)));
+            console.log(new Float32Array(buffer.slice(13, 17)));
         });
 };
